[refs #85843] Fix circular import
